LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.font.default-expected.txt

2d.text.font.default


FAIL OffscreenCanvas test: 2d.text.font.default assert_equals: ctx.font === '10px sans-serif' (got [undefined], expected 10px sans-serif[string]) expected (string) "10px sans-serif" but got (undefined) undefined
